start: doors event
