+++ Makefile
@@ -6,7 +6,7 @@
 #    By: decordel <decordel@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/06/27 21:18:07 by decordel          #+#    #+#              #
-#    Updated: 2022/07/13 19:19:54 by decordel         ###   ########.fr        #
+#    Updated: 2022/07/14 03:07:12 by decordel         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -21,6 +21,7 @@ SRCS	=	main.c				\
 			mlx_init.c			\
 			player.c			\
 			player_controller.c	\
+			player_move.c		\
 			draw.c				\
 			walls.c				\
 			raycast.c			\
